REF: test_append, parametrize, move misplced
