+++ pandas/tests/frame/test_analytics.py
@@ -1248,15 +1248,6 @@ class TestDataFrameAnalytics:
     # ---------------------------------------------------------------------
     # Unsorted
 
-    def test_series_nat_conversion(self):
-        # GH 18521
-        # Check rank does not mutate DataFrame
-        df = DataFrame(np.random.randn(10, 3), dtype="float64")
-        expected = df.copy()
-        df.rank()
-        result = df
-        tm.assert_frame_equal(result, expected)
-
     def test_series_broadcasting(self):
         # smoke test for numpy warnings
         # GH 16378, GH 16306
